video_explicit/requirements.txt
opencv-python>=4.9.0.80
ffmpeg-python>=0.2.0
easyocr==1.7.1
numpy>=1.26.4
torch==2.2.2
scipy>=1.12.0
